Removed a wide card from help page
Removed a wide card from the PHP file as it conatined text that was a F&A question from the design and wont be needed for the demo. Created another class in the "style.css" file for the help page. This will be used for the hidden class that the span tag will use for the text that will appear when the text within one of the wide cards is selected. This will create a drop down effect. This class is called "caret-text".