Adding collapsible JS
Removed orginal way off using event listener methods for the caret drop down text. For trial and error reasons I have tested a new way for the first title in the first top wide card. Have made the title a button so that when the button is cliked on the DIV class "caret-content" will be revealed containg the span tage text which is the answer text.