Solved JS collapsible issue
Solved issue with making collapsible work for each of the title question drop downs. Within the JS file i forgot to include "document.addEventListener("DOMContentLoaded", () => {" at the start of the script. Will now apply to the other six wild cards on the page by implementing the button tag for the question titles.