Added caret icon svg
Have added caret icon svg to each title wide card. Have made two classes in CSS to style the svg and change the behaviour of it's rotation when the class is toggled active in JS. In the process of finishing the JS code for the caret icon to turn pointing downwards when each drop down is clicked on.