Made final adjustments to JS code and other content
Have made final adjustments to JS code for the collaspible drop down text on the help page. Was missing this suitable line of code "this.getElementsByClassName("caret-icon")[0].classList.toggle("active");" that toggles the active.caret class to turn on, causing the caert svg to tun 90 degrees downward on click drop down with the text. Adjusted the bordering of the drop down text better. Presentation now looks more in proportion.