Fixed an issue where the Charts page would fetch data too quickly.
When recording falls, it can take a second for the API to process the data and execute the appropriate SQL queries to store the data. As such, the Charts page now waits before fetching data from the API, but only if the previous page the user was on was the Falls page.