Have started adding content into style.css file for web page.
Have created a CSS class named "view-small-card" in style.css file. Each small card border on the page will store the characters of info for each field from a selected user when a profile is being viewed. For example, the Username & First Name. Have set the width calculation to 25% so four wide cards can fit on the web page per row. The api end points will pull out the info and will store them in each of the small cards.