Created "Export Data" button. Have also put a div class around the buttons named "button-wrapper".
Created and styled HTML button from CSS class "page-button". This button will be given functionality in the future whereby when clicked on, the users info being displayed in the rows of small cards will be exported to a CSV file format & can be saved. "button-wrapper" class has also placed buttons in the centre below the small cards.