Have given each small card an internal Span Class & ID.
Each span class which is the "field-title" class will make the title of each data field name bold. The span id name relates to the getElementById JS value so that when the id is used and referenced to a used variable: the span text will be generated from the value of that field taken from the api.