Have started to create if statement inside event listener.
Have started to create if statement inside event listener to start pulling in the data field values for their corresponding small card containers. Using a XML HTTP request, a XHR .open element was used containing the relevant "read" api endpoint. This is so the users info can read in using the GET request .