Have finished adding document.getElementByIds for user fields in JS file.
Have finished adding document.getElementById's with JS variables for api data fields to be pulled in from the relevant endpoint for the user being currently viewed and then placed into the corresponding span tag within each small card. The variables outside of the if statement get the span ids to then be assigned to the card variable name. Within the if statement, another 12 variables are created which equal to their relevant api fields from the "read" endpoint. Finally, the 12 card variables then equal to the 12 endpoint variables and use ".innerText" to place the data value of what that field is into the HTML span on the web page. Will code comment and tidy up code.