JS Dark theme IF statement
Have included the universal Dark theme if statement within the main DOM loaded function for script. Checks to see if localStorage has the key:'theme', and if it does is, the value will become value:'dark'. When that is true, it sets the body with an attribute to turn the theme dark.