Have created "index.js" file for landing page (index.php)
This is so it can store the Dark theme JS code for the web page. DOM loaded function checks to see if localStorage has the key:'theme', and if it does it will be set to value:'dark'. When that is true, it sets the body with an attribute to turn the theme dark. Have also linked JS file to web page.