Have included Dark theme JS code for developers page.
DOM loaded function checks to see if localStorage has the key:'theme', and if it does it will be set to value:'dark'. When that is true, it sets the body with an attribute to turn the theme dark. Have created JS file named "developers.js" for this and have linked it to "developers.php".