Started modifying Resize CSS for Landing page (jndex.php)
Have started to add resize override properties in the resize.css file based off of classes from style.css. This is so that items such as the HTML buttons within the nav bar div do not go onto a sperate line when browser window width goes below 1000px when width is being shortened. Not finished yet.