Inserted whole JS code into the VerifySession function.
This is so that when an admin user is currently logged in with their admin credentials, the session token that contains the details for the logged in user of the web page (whole site) can be recognised. The VerifySession function stems from the "global.js" file. Web page has also been linked to this JS file before any other script within the <head> tag in "index.php".