Included Edit and Delete button event listener functions.
Included Edit and Delete button event listener functions for both question types. Only editable questions that can be viewed can be edited or deleted. The delete button will remove a question card div and it's data from the api endpoint. The edit button on a question card will take the admin to the edit questions page for that question ID.